Popup menus in "Language Info" panel
A lot of very clunky code but it does most of what it needs to,
except for that it doesn’t save the changes if you rename anything
other than a language, but I was thinking I don’t really need to make
this work as you will have to retool this sort of thing for your
database structure anyway, rather than use the current placeholder
dictionary system that I put in place?

It took me 3 hours to write this - I thought it would take 20 mins! I’m
going to have to get on to important other stuff like making sure that
the data table is appropriate :S